WNM.UI: Turn on LEDs by row (failed experiment)
...instead of one by one.

The idea was to have less interrupts by handling all the LEDs in the
same raw at the same time.

In practice,powering multiple LEDs at the same time is not great as they
will not have the same brightness, especially when the LEDs are not of
the same color.

I put this patch in to log the experiment.